+++ src/libc/fd/ofdtx_pwrite.c
@@ -250,7 +250,7 @@ ofdtx_pwrite_any_undo(void)
 int
 ofdtx_pwrite_undo(struct ofdtx *ofdtx, int fildes, int cookie)
 {
-    static ssize_t (* const pwrite_undo[][4])(void) = {
+    static int (* const pwrite_undo[][4])(void) = {
         {NULL, NULL,                  NULL,                  NULL},
         {NULL, ofdtx_pwrite_any_undo, ofdtx_pwrite_any_undo, NULL},
         {NULL, NULL,                  NULL,                  NULL},
